moved the Header and rating files to a different components
