+++ src/App.tsx
@@ -1,5 +1,5 @@
 import "./App.scss";
-import Header from "./components/Header";
+import Header from "./components/Header/Header";
 import Hero from "./components/Hero";
 import Popular from "./components/Popular";
